Add ColorPalette service
A color palette is composed of two colors and the N previous selected
colors.  In our case, N is set to 10 as asked by the client.

A color is simply a 32-bits value in the RGBA format, see
<https://en.wikipedia.org/wiki/RGBA_color_space>.

The _primary_ and _secondary_ colors can be swapped, without changing
the previous selected colors.

Selection of a new color result in changing either the primary or the
secondary color.  The color that was in the position of
primary/secondary is then add into the previous selected colors
history, using a ring buffer.

Signed-off-by: Olivier Dion <olivier.dion@polymtl.ca>
Signed-off-by: Tristan Philis <tristan.philis@polymtl.ca>
Reviewed-by: Bei Ning Pan <bei-nin.pan@polymtl.ca>