ToolService injects bucket and svg services
The tool service should _not_ create tool object.  They should be
service that are injected.  The ToolService job is to _select_ the
current tool and notify the SVGService.

Adding new tools is simply a matter of injecting its service into the
ToolService and adding its selector into the ToolComponent selector.

Signed-off-by: Olivier Dion <olivier.dion@polymtl.ca>