Add a color picker component
To be implemented for new drawing form and toolbar

* makeColorPallette

The current gradient is linear (canvas): colors horizontally and transparency
vertically.

* listenToColor

I create mouseDown, mouseUp and mouseMove observables. The color picker
subscribes to mouseDown and then subscribes to mouseMove. When mouseUp
is triggered, it unsubscribes to mouseMove.

* getColor

The color picker extracts the color on the current pixel which the
mouse is positionned on.

Signed-off-by: Simon Berhanu <simon1515@hotmail.com>