Fix #117 - Infinite new draw pop up when C-o
See <https://github.com/oldiob/LOG2990/issues/117>.

In order to call the dialog only one time, we need to keep a boolean
flag (isDialogOpened) to keep track of the state.  Once the dialog is
closed, we reset this flag.

For performance reasons, the callback should be called on keyup and
not keydown.  This way, if the user keep C-o pushed, it won't call
anything.

Also, since keyup is _after_ keydown, and since C-o is binded to open
file by default after keydown, we need to prevent the default action.
This is done by the preventDefault method.

Signed-off-by: Olivier Dion <olivier.dion@polymtl.ca>