Add selector tool
TODO remove the width.  The width is here only so I the tool can be
tested with the pen, etc.

Would be nice if we can also move the diff and union method to
utils/set.

The boxElement is the dashed selection.  The previewElement is the
actual selection appearing on the screen.  This might be confusing and
could be renamed in the future.

The policy dictated on how to merge the selected object and the
current selection.

A policy of false means, overwrite the selected object with the
selection.  A policy of true means, take the symetric difference of
the selected objects and the selection.  See
SelectorTool::computeSelection.

Signed-off-by: Olivier Dion <olivier.dion@polymtl.ca>